Refactor some missed folders
