+++ packages/yoastseo/spec/parsedPaper/structure/tree/TextContainerSpec.js
@@ -1,4 +1,4 @@
-import TextContainer from "../../../src/parsedPaper/structure/tree/TextContainer";
+import TextContainer from "../../../../src/parsedPaper/structure/tree/TextContainer";
 
 describe( "TextContainer", () => {
 	describe( "constructor", () => {
