+++ packages/yoastseo/spec/parsedPaper/structure/tree/nodes/ParagraphSpec.js
@@ -1,4 +1,4 @@
-import Paragraph from "../../../../src/parsedPaper/structure/tree/nodes/Paragraph";
+import Paragraph from "../../../../../src/parsedPaper/structure/tree/nodes/Paragraph";
 
 describe( "Paragraph tree node", () => {
 	describe( "constructor", () => {
